More tests, some refactoring

+++ go/libraries/doltcore/sqle/enginetest/dolt_queries_schema_merge.go
@@ -43,7 +43,7 @@ var SchemaChangeTestsForDataConflicts = []MergeScriptTest{
 		Assertions: []queries.ScriptTestAssertion{
 			{
 				Query:    "select * from dolt_preview_merge_conflicts_summary('main', 'right');",
-				Expected: []sql.Row{{"t", uint(1), uint(0)}},
+				Expected: []sql.Row{{"t", uint64(1), uint64(0)}},
 			},
 			{
 				Query:    "call dolt_merge('right');",
